requirements.txt
# Runtime requirements

uvicorn==0.32.0
fastapi==0.115.3
pyyaml==6.0.2
rich==13.9.3
azure-identity==1.19.0
fastapi_azure_auth==5.0.1
